Use long, not int

+++ src/csharp/Intel/Iced/Intel/Assembler/AssemblerMemoryOperand.cs
@@ -159,7 +159,7 @@ namespace Iced.Intel {
 		/// <param name="left">The memory operand.</param>
 		/// <param name="displacement">displacement.</param>
 		/// <returns></returns>
-		public static AssemblerMemoryOperand operator +(AssemblerMemoryOperand left, int displacement) =>
+		public static AssemblerMemoryOperand operator +(AssemblerMemoryOperand left, long displacement) =>
 			new AssemblerMemoryOperand(left.Size, Register.None, left.Base, left.Index, left.Scale, left.Displacement + displacement, left.Flags);
 
 		/// <summary>
@@ -168,7 +168,7 @@ namespace Iced.Intel {
 		/// <param name="left">The memory operand.</param>
 		/// <param name="displacement">displacement.</param>
 		/// <returns></returns>
-		public static AssemblerMemoryOperand operator -(AssemblerMemoryOperand left, int displacement) =>
+		public static AssemblerMemoryOperand operator -(AssemblerMemoryOperand left, long displacement) =>
 			new AssemblerMemoryOperand(left.Size, Register.None, left.Base, left.Index, left.Scale, left.Displacement - displacement, left.Flags);
 
 		/// <summary>
